Change count stat variable type from int to long (#35)
Change count stat variable type from int to long

+++ src/main/java/io/tehuti/metrics/stats/Count.java
@@ -20,13 +20,13 @@ import io.tehuti.metrics.MetricConfig;
  */
 public class Count implements MeasurableStat {
 
-    private int count;
+    private long count;
 
     public Count() {
         this.count = 0;
     }
 
-    public Count(int value) {
+    public Count(long value) {
         this.count = value;
     }
